Rename `commands` to `cmd`
This rename bring the packages more in line with how the rest of the
community names their folders
