+++ cmd/awskms/common.go
