+++ cmd/awskms/multiregionkey.go
