+++ cmd/export.go
@@ -1,4 +1,4 @@
-package commands
+package cmd
 
 import (
 	"context"
