+++ main.go
@@ -13,8 +13,8 @@ import (
 	"github.com/dcoker/biscuit/algorithms/aesgcm256"
 	"github.com/dcoker/biscuit/algorithms/plain"
 	"github.com/dcoker/biscuit/algorithms/secretbox"
-	"github.com/dcoker/biscuit/commands"
-	"github.com/dcoker/biscuit/commands/awskms"
+	"github.com/dcoker/biscuit/cmd"
+	"github.com/dcoker/biscuit/cmd/awskms"
 	"github.com/dcoker/biscuit/shared"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
@@ -61,10 +61,10 @@ func main() {
 	kmsGrantsCreateFlags := kmsGrantsFlags.Command("create", mustAsset("data/kmsgrantcreate.txt"))
 	kmsGrantsRetireFlags := kmsGrantsFlags.Command("retire", mustAsset("data/kmsgrantsretire.txt"))
 
-	getCommand := commands.NewGet(getFlags)
-	writeCommand := commands.NewPut(putFlags)
-	listCommand := commands.NewList(listFlags)
-	exportCommand := commands.NewExport(exportFlags)
+	getCommand := cmd.NewGet(getFlags)
+	writeCommand := cmd.NewPut(putFlags)
+	listCommand := cmd.NewList(listFlags)
+	exportCommand := cmd.NewExport(exportFlags)
 	kmsIDCommand := awskms.KmsGetCallerIdentity{}
 	kmsEditKeyPolicy := awskms.NewKmsEditKeyPolicy(kmsEditKeyPolicyFlags)
 	kmsGrantsListCommand := awskms.NewKmsGrantsList(kmsGrantsListFlags)
